working on database files

+++ src/db/table_templates.sql
@@ -1,7 +1,11 @@
+-- noinspection SqlDialectInspectionForFile
+
+-- noinspection SqlNoDataSourceInspectionForFile
+
 CREATE TABLE IF NOT EXISTS votes (
     id          integer     PRIMARY KEY,
     uid         integer     NOT NULL,
-    timestamp   integer     NOT NULL,
+    timestamp   float     NOT NULL,
     score       integer     NOT NULL check(score between 0 and 5),
     latitude    real        check(latitude between -90 and 90),
     longitude   real        check(longitude between -180 and 180),
@@ -17,7 +21,7 @@ CREATE TABLE IF NOT EXISTS posts (
     message     varchar     NOT NULL,
     upvotes     integer     NOT NULL check(upvotes >= 0),
     downvotes   integer     NOT NULL check(downvotes >= 0),
-    timestamp   integer     NOT NULL,
+    timestamp   float     NOT NULL,
     latitude    real        check(latitude between -90 and 90),
     longitude   real        check(longitude between -180 and 180),
     logical_loc varchar    --TODO: restrict this to simple formats?
